Import of all old definitions from iamcompact-nomenclature.
This commit imports all the definitions that were present in the main
iamcompact-nomenclature repository as of 2024-10-14. This is part of
refactoring of definitions into a separate repository, and separating the
definitions into IAM COMPACT-specific names and overrides, and names that are
shared with the common-definitions repository.

Many of the files here will be removed in future commits, both to conform to
the eventual official list of IAM COMPACT variable names, and to ensure that
the variable definitions here are additional to or overrides of
common-definitions, not duplicates.