Adding all energy yaml files from common-definitions to common-definitions-tags folder.
Need to duplicate most of it to use as weights.

This is kind of messy, and will require work to keep up to date. Maybe an
indication that we should instead go for a single repo and merge in changes
happening in common-definitions. Perhaps with a separate subfolder tree for
files and folders that are added in IAM COMPACT, named `iam-compact-added` or
something like that. But no time to do that now.